Broaden audit report detection
